% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{chatgpt_addin}
\alias{chatgpt_addin}
\title{ChatGPT Addin}
\usage{
chatgpt_addin()
}
\value{
The function does not return a value. It runs the Shiny gadget within the RStudio
  environment, allowing the user to interact with the ChatGPT model.
}
\description{
This function launches an interactive Shiny gadget that serves as an RStudio addin
for interacting with OpenAI's GPT model. The addin
provides various functionality, such as asking ChatGPT, commenting or explaining
selected code, creating unit tests, optimizing or refactoring selected code, and more.
}
